Struct viewdock::Split

pub struct Split {
    pub left: Option<SplitHandle>,
    pub right: Option<SplitHandle>,
    pub left_docks: Container,
    pub right_docks: Container,
    pub ratio: f32,
    pub direction: Direction,
    pub handle: SplitHandle,
    // some fields omitted
}

Holds data of the splits and containers. There are always two containers in a split if it has no split childern. If the Split has another split in left it's expected that the left_docks container will be empty as they docks would be on a lower level.

Fields
left: Option<SplitHandle>

left/top split

right: Option<SplitHandle>

right/bottom split

left_docks: Container

left/top docks

right_docks: Container

right/top docks

ratio: f32

ratio value of how much of each side that is visible. 1.0 = right/bottom fully visible

direction: Direction

Direction of the split

handle: SplitHandle

Handle of the spliter

Methods

impl Split

fn new(direction: Direction, ratio: f32, handle: SplitHandle) -> Split

fn is_left_zero_and_none(&self) -> bool

fn is_right_zero_and_none(&self) -> bool

fn no_split(&mut self, direction: Direction, dock_handle: DockHandle) -> bool

fn split_left(&mut self, split_handle: SplitHandle, dock_handle: DockHandle, direction: Direction) -> Option<Split>

fn split_right(&mut self, split_handle: SplitHandle, dock_handle: DockHandle, direction: Direction) -> Option<Split>

fn get_sizer_from_rect_horizontal(rect: Rect, size: f32) -> Rect

fn get_sizer_from_rect_vertical(rect: Rect, size: f32) -> Rect

fn is_hovering_rect(&self, pos: (f32, f32), border_size: f32, rect: Rect) -> bool

fn map_rect_to_delta(&self, delta: (f32, f32)) -> f32

fn change_ratio(&mut self, delta: (f32, f32))

fn dump_info(&self, level: i32)