Struct viewdock::Workspace

pub struct Workspace {
    pub splits: Vec<Split>,
    pub rect: Rect,
    pub window_border: f32,
    // some fields omitted
}

Top level structure that holds an array of all the splits and the rect size of of the full layout. This size is then propagated downwards and recalculated depending on the tree structure.

Fields
splits: Vec<Split> rect: Rect window_border: f32

border size of the windows (in pixels)

Methods

impl Workspace

fn new(rect: Rect) -> Result<Workspace>

Construct a new workspace. The rect has to be y >= 0, x >= 0, width > 0 and height > 0

fn calc_horizontal_sizing(rect: Rect, ratio: f32) -> (Rect, Rect)

fn calc_vertical_sizing(rect: Rect, ratio: f32) -> (Rect, Rect)

fn split(&mut self, dock_handle: DockHandle, direction: Direction)

fn split_by_dock_handle(&mut self, direction: Direction, find_handle: DockHandle, handle: DockHandle)

fn new_split(&mut self, handle: DockHandle, direction: Direction)

fn get_rect_by_handle(&self, handle: DockHandle) -> Option<Rect>

fn get_hover_dock(&self, pos: (f32, f32)) -> Option<DockHandle>

fn update(&mut self)

fn drag_sizer(&mut self, handle: SplitHandle, delta: (f32, f32))

fn is_hovering_sizer(&self, pos: (f32, f32)) -> Option<(SplitHandle, Direction)>

fn dump_tree(&self)

fn dump_tree_linear(&self)

fn delete_by_handle(&mut self, handle: DockHandle)

fn get_docks(&self) -> Vec<Dock>

fn save(&self, file_name: &str) -> Result<()>

fn load(file_name: &str) -> Workspace